requirements.txt
asgiref==3.8.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.0
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
idna==3.11
pillow==10.4.0
PyJWT==2.0.0
pytz==2025.2
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.2.3
